restructure & clenaup
