+++ test/e2e/target-cleanup/setup_test.go
@@ -2,6 +2,7 @@ package targetcleanuptest
 
 import (
 	"io/ioutil"
+	"os"
 
 	"github.com/benchkram/bob/bob"
 	"github.com/benchkram/bob/pkg/nix"
@@ -34,10 +35,6 @@ func bobSetup(opts ...bob.Option) (_ *bob.B, err error) {
 	)
 }
 
-// tmpFiles tracks temporarily created files in these tests
-// to be cleaned up at the end.
-var tmpFiles []string
-
 func NixBuilder() (*bob.NixBuilder, error) {
 	file, err := ioutil.TempFile("", ".nix_cache*")
 	if err != nil {
@@ -54,4 +51,28 @@ func NixBuilder() (*bob.NixBuilder, error) {
 	}
 
 	return bob.NewNixBuilder(bob.WithCache(cache)), nil
+}
+
+// useBobfile sets the right bobfile to be used for test
+func useBobfile(name string) error {
+	return os.Rename(name+".yaml", "bob.yaml")
+}
+
+// releaseBobfile will revert changes done in useBobfile
+func releaseBobfile(name string) error {
+	return os.Rename("bob.yaml", name+".yaml")
+}
+
+// readDir returns the dir entries as string array
+func readDir(dir string) ([]string, error) {
+	entries, err := os.ReadDir(dir)
+	if err != nil {
+		return []string{}, err
+	}
+
+	var contents []string
+	for _, e := range entries {
+		contents = append(contents, e.Name())
+	}
+	return contents, nil
 }
