+++ test/e2e/target-cleanup/with_dir_target/bob.yaml
